+++ src/Umbraco.Tests/LegacyXmlPublishedCache/DomainCache.cs
@@ -1,11 +1,12 @@
 ﻿using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using Umbraco.Web.Routing;
 using Umbraco.Core;
 using Umbraco.Core.Services;
+using Umbraco.Web.PublishedCache;
+using Umbraco.Web.Routing;
 
-namespace Umbraco.Web.PublishedCache.XmlPublishedCache
+namespace Umbraco.Tests.LegacyXmlPublishedCache
 {
     internal class DomainCache : IDomainCache
     {
