+++ src/Umbraco.Tests/LegacyXmlPublishedCache/PublishedSnapshot.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using Umbraco.Core;
 using Umbraco.Core.Cache;
+using Umbraco.Web.PublishedCache;
 
-namespace Umbraco.Web.PublishedCache.XmlPublishedCache
+namespace Umbraco.Tests.LegacyXmlPublishedCache
 {
     /// <summary>
     /// Implements a published snapshot.
