+++ src/Umbraco.Tests/LegacyXmlPublishedCache/XmlPublishedProperty.cs
@@ -5,7 +5,7 @@ using Umbraco.Core.Models.PublishedContent;
 using Umbraco.Core.PropertyEditors;
 using Umbraco.Core.Xml;
 
-namespace Umbraco.Web.PublishedCache.XmlPublishedCache
+namespace Umbraco.Tests.LegacyXmlPublishedCache
 {
 
     /// <summary>
