--- src/Umbraco.Web.UI.Client/src/views/dashboard/settings/xmldataintegrityreport.html
@@ -1,29 +1,0 @@
-<div id="examineManagement" ng-controller="Umbraco.Dashboard.XmlDataIntegrityReportController">
-
-    <h3>Xml Cache Data integrity</h3>
-
-    <div ng-show="loading">
-        Loading...
-    </div>
-
-    <p>
-        This checks the data integrity for the xml structures for content, media and members that are stored in the cmsContentXml table.
-        This does not check the data integrity of the xml cache file, only the xml structures stored in the database used to create the xml cache file.
-    </p>
-    <div ng-repeat="(key, value) in items" style="padding: 4px 0;">
-        <span ><strong>{{value.label}}&nbsp;...</strong></span>
-        <span ng-if="value.checking" class="text-info">Checking...</span>
-        <span ng-if="!value.checking && !value.invalid" class="text-success"><strong>Ok</strong></span>
-        <span ng-if="!value.checking && value.invalid" class="text-error"><strong>Error</strong></span>
-        <div ng-show="value.fix && !value.fixing">
-            <button type="button" ng-click="fix(value)" class="btn btn-danger" ng-if="value.invalid">
-                <span>Fix</span>
-
-            </button>
-        </div>
-        <div class="umb-loader-wrapper" ng-show="value.fixing">
-            <div class="umb-loader"></div>
-        </div>
-    </div>
-
-</div>
